+++ NEWS.md
@@ -1,6 +1,8 @@
 # BayesSurvive 0.0.2
 
 * Add vignette
+* Add function `coef.BayesSurvive()` to extract estimatiated coefficients and uncertainty
+* Add function `VS()` to determine variable selection
 * Fix documentation issues
 
 # BayesSurvive 0.0.1